ComunioApp2/requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1646806365504/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
astor==0.8.1
astunparse==1.6.3
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
blinker==1.4
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028898966/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646056590078/work
coreapi==2.3.3
coreschema==0.0.4
cryptography @ file:///tmp/build/80754af9/cryptography_1633520369886/work
defusedxml==0.7.1
diff-match-patch==20200713
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-coreapi==2.3.0
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
flatbuffers @ file:///home/ktietz/cip/python-flatbuffers_1634039120618/work
frozenlist @ file:///tmp/build/80754af9/frozenlist_1637767113340/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///tmp/build/80754af9/grpcio_1637590821884/work
h5py @ file:///tmp/build/80754af9/h5py_1608649014695/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
import-export==0.2.67.dev6
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648544546694/work
itypes==1.2.0
Jinja2==3.1.2
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
Markdown @ file:///tmp/build/80754af9/markdown_1614363852612/work
MarkupPy==1.14
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
multidict @ file:///opt/conda/conda-bld/multidict_1640703752579/work
numexpr @ file:///tmp/build/80754af9/numexpr_1640689833592/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1651563629415/work
oauthlib @ file:///opt/conda/conda-bld/oauthlib_1644332107998/work
odfpy==1.4.1
openpyxl==3.0.9
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.2
protobuf==3.19.1
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1612298595717/work
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///tmp/build/80754af9/pyjwt_1619682484438/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1642617106979/work
scipy @ file:///tmp/build/80754af9/scipy_1641555004408/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tablib==3.2.1
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_cpu/tensorflow-base_1614360980945/work/tensorflow_pkg/tensorflow-2.4.1-cp39-cp39-linux_x86_64.whl
tensorflow-estimator @ file:///home/builder/adipietro/tf/tensorflow-estimator_1630508970172/work/tensorflow_estimator-2.6.0-py2.py3-none-any.whl
termcolor==1.1.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
uritemplate==4.1.1
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
wrapt @ file:///tmp/build/80754af9/wrapt_1638433857881/work
xlrd==2.0.1
xlwt==1.3.0
yarl @ file:///tmp/build/80754af9/yarl_1606939947528/work
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
